Rename wanteditem to item and move to core module
