+++ autosubliminal/subchecker.py
@@ -20,13 +20,13 @@ from subliminal.video import Episode, Movie
 
 import autosubliminal
 from autosubliminal import utils
+from autosubliminal.core.item import WantedItem
 from autosubliminal.core.scheduler import ScheduledProcess
 from autosubliminal.db import WantedItems
 from autosubliminal.postprocessor import PostProcessor
 from autosubliminal.providers import provider_cache
 from autosubliminal.providers.addic7ed import Addic7edSubtitle as Addic7edSubtitleRandomUserAgent
 from autosubliminal.subdownloader import SubDownloader
-from autosubliminal.wanteditem import WantedItem
 
 log = logging.getLogger(__name__)
