+++ web/templates/home/inc_home-wanteditems.mako
@@ -2,8 +2,8 @@
     import os
 
     import autosubliminal
+    from autosubliminal.core.item import WantedItem
     from autosubliminal.utils import convert_timestamp, display_item, display_title, get_file_size, count_wanted_items
-    from autosubliminal.wanteditem import WantedItem
 %>
 
 <%block name="bodyContent">
